+++ CHANGELOG.md
@@ -2,6 +2,18 @@
 
 All notable changes to `company-info` will be documented in this file.
 
+## 1.1.1 - 2022-03-29
+
+- Fixes for CompanyInfo::fake().
+
+## 1.1.0 - 2022-03-29
+
+- Implemented CompanyInfo::fake().
+
+## 1.0.1 - 2022-03-28
+
+- Fixed phpdoc return type of the lookupName/Number functions.
+
 ## 1.0.0 - 2022-03-25
 
 - Initial release.